New version * eliminated need for Python 2, everything runs with Python3.5 now * eliminated need for Picrust2 install, it is now used as a local module
This depends on a softlink between the picrust2/scripts folder and the
 picrust2/picrust2 folder. an update could potentially det rid of the
 separation entirely,
 but would require updating all the import calls in the scripts we use..

* The pipeline is now completely path agnostic.
It can be called using the FANCy-otu and FANCy-bam scripts from anywhere
and the output folders (log,output,bamdir/otudir,data) are created where
the user wants them, and the data written there.

This should make it more easy to use.
* The aforementioned FANCy-otu.sh and FANCy-bam.sh scripts were created
to faciliate calling the pipeline for users.

here's an example:

./FANCy-otu.sh otuTable.txt otuSeqs.fasta matchfiles log out meta.csv /
                                          0.05 tss yes no