Added OTU dataset
This OTU dataset, composed of the compressed otu Fasta sequences, the ptu table quantifying the otu's present in each sample group, and the sample<->condition correspondence table, were generated using the same artificial data as the BAM files.

To use in the pipeline, decompress the zip file. the other files can be used as is, with "yes" and "no" as the sample conditions to be compared.